[feat] 스킬 기능 기본 구현
[refactor] 강화 셀 중복 최소화
