--- Assets/Scripts/Cell/Mod.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 45e7041b33ed8b748b8fdbe227557cf1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
