--- Assets/Scripts/Cell/Mod/AttackDataCell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd59e94a551cd1847bad9e6713c34bec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
