--- Assets/Scripts/Cell/Mod/MaxHpDataCell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c888b9206e2459645943c6dbb5b84267
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
