+++ Assets/Scripts/Cell/Player/DataCell.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c3e4eec6505bed43b5e9fe70177c62c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
