+++ Assets/Scripts/Cell/Player/SpeedDataCell.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0018eda8629831d4fbd44701639eeb07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
